Rename bluealsa-cli tool to bluealsactl
