+++ NEWS
@@ -1,6 +1,8 @@
 unreleased
 ==========
 
+- renamed bluealsa-cli to bluealsactl (no backward compatibility)
+
 bluez-alsa v4.3.1 (2024-08-30)
 ==============================
 
@@ -40,7 +42,7 @@ bluez-alsa v4.1.1 (2023-06-24)
 bluez-alsa v4.1.0 (2023-05-23)
 ==============================
 
-- removed deprecated org.bluealsa.Manger1 D-Bus interface
+- removed deprecated org.bluealsa.Manager1 D-Bus interface
 - persistent storage for PCM volume and mute state
 - PCM volume control with oFono HFP-AG and HFP-HF profiles
 - transport running state exported in PCM D-Bus interface
@@ -59,7 +61,7 @@ bluez-alsa v4.0.0 (2022-06-03)
 ==============================
 
 - changes in command line options (no backward compatibility)
-- implement D-Bus ObjectManager (deprecate org.bluealsa.Manger1)
+- implement D-Bus ObjectManager (deprecate org.bluealsa.Manager1)
 - optional support for A2DP FastStream codec (music & voice)
 - optional support for A2DP LC3plus codec (music & voice)
 - enhanced SBC XQ mode (XQ+) which bumps bitrate to 595 kbps
