+++ src/asound/bluealsa-pcm.c
@@ -1,5 +1,5 @@
 /*
- * bluealsa-pcm.c
+ * BlueALSA - asound/bluealsa-pcm.c
  * Copyright (c) 2016-2024 Arkadiusz Bokowy
  *
  * This file is a part of bluez-alsa.
