+++ src/bluealsactl/cmd-list-services.c
@@ -1,5 +1,5 @@
 /*
- * BlueALSA - cmd-list-services.c
+ * BlueALSA - bluealsactl/cmd-list-services.c
  * Copyright (c) 2016-2024 Arkadiusz Bokowy
  *
  * This file is a part of bluez-alsa.
@@ -16,7 +16,7 @@
 
 #include <dbus/dbus.h>
 
-#include "cli.h"
+#include "bluealsactl.h"
 #include "shared/dbus-client.h"
 
 static bool print_service(const char *name, void *data) {
@@ -28,7 +28,7 @@ static bool print_service(const char *name, void *data) {
 
 static void usage(const char *command) {
 	printf("List all BlueALSA services.\n\n");
-	cli_print_usage("%s [OPTION]...", command);
+	bactl_print_usage("%s [OPTION]...", command);
 	printf("\nOptions:\n"
 			"  -h, --help\t\tShow this message and exit\n"
 	);
@@ -47,7 +47,7 @@ static int cmd_list_services_func(int argc, char *argv[]) {
 
 	opterr = 0;
 	while ((opt = getopt_long(argc, argv, opts, longopts, NULL)) != -1) {
-		if (cli_parse_common_options(opt))
+		if (bactl_parse_common_options(opt))
 			continue;
 		switch (opt) {
 		case 'h' /* --help */ :
@@ -65,7 +65,7 @@ static int cmd_list_services_func(int argc, char *argv[]) {
 	}
 
 	DBusError err = DBUS_ERROR_INIT;
-	cli_get_ba_services(print_service, NULL, &err);
+	bactl_get_ba_services(print_service, NULL, &err);
 	if (dbus_error_is_set(&err)) {
 		cmd_print_error("D-Bus error: %s", err.message);
 		return EXIT_FAILURE;
@@ -74,7 +74,7 @@ static int cmd_list_services_func(int argc, char *argv[]) {
 	return EXIT_SUCCESS;
 }
 
-const struct cli_command cmd_list_services = {
+const struct bactl_command cmd_list_services = {
 	"list-services",
 	"List all BlueALSA services",
 	cmd_list_services_func,
