+++ src/bluealsactl/cmd-mute.c
@@ -1,5 +1,5 @@
 /*
- * BlueALSA - cmd-delay-adjustment.c
+ * BlueALSA - bluealsactl/cmd-mute.c
  * Copyright (c) 2016-2024 Arkadiusz Bokowy
  *
  * This file is a part of bluez-alsa.
@@ -8,33 +8,31 @@
  *
  */
 
-#include <errno.h>
 #include <getopt.h>
-#include <math.h>
-#include <stdint.h>
+#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 
 #include <dbus/dbus.h>
 
-#include "cli.h"
+#include "bluealsactl.h"
 #include "shared/dbus-client-pcm.h"
 
 static void usage(const char *command) {
-	printf("Get or set the delay adjustment of the given PCM.\n\n");
-	cli_print_usage("%s [OPTION]... PCM-PATH [ADJUSTMENT]", command);
+	printf("Get or set the mute switch of the given PCM.\n\n");
+	bactl_print_usage("%s [OPTION]... PCM-PATH [STATE [STATE]]", command);
 	printf("\nOptions:\n"
 			"  -h, --help\t\tShow this message and exit\n"
 			"\nPositional arguments:\n"
 			"  PCM-PATH\tBlueALSA PCM D-Bus object path\n"
-			"  ADJUSTMENT\tAdjustment value (+/-), in milliseconds\n"
+			"  STATE\t\tEnable or disable mute switch\n"
 	);
 }
 
-static int cmd_delay_adjustment_func(int argc, char *argv[]) {
+static int cmd_mute_func(int argc, char *argv[]) {
 
 	int opt;
-	const char *opts = "+hqv";
+	const char *opts = "hqv";
 	const struct option longopts[] = {
 		{ "help", no_argument, NULL, 'h' },
 		{ "quiet", no_argument, NULL, 'q' },
@@ -44,11 +42,15 @@ static int cmd_delay_adjustment_func(int argc, char *argv[]) {
 
 	opterr = 0;
 	while ((opt = getopt_long(argc, argv, opts, longopts, NULL)) != -1) {
-		if (cli_parse_common_options(opt))
+		if (bactl_parse_common_options(opt))
 			continue;
-		if (opt == 'h') { /* --help */
+		switch (opt) {
+		case 'h' /* --help */ :
 			usage(argv[0]);
 			return EXIT_SUCCESS;
+		default:
+			cmd_print_error("Invalid argument '%s'", argv[optind - 1]);
+			return EXIT_FAILURE;
 		}
 	}
 
@@ -56,7 +58,7 @@ static int cmd_delay_adjustment_func(int argc, char *argv[]) {
 		cmd_print_error("Missing BlueALSA PCM path argument");
 		return EXIT_FAILURE;
 	}
-	if (argc - optind > 2) {
+	if (argc - optind > 3) {
 		cmd_print_error("Invalid number of arguments");
 		return EXIT_FAILURE;
 	}
@@ -65,42 +67,50 @@ static int cmd_delay_adjustment_func(int argc, char *argv[]) {
 	const char *path = argv[optind];
 
 	struct ba_pcm pcm;
-	if (!cli_get_ba_pcm(path, &pcm, &err)) {
+	if (!bactl_get_ba_pcm(path, &pcm, &err)) {
 		cmd_print_error("Couldn't get BlueALSA PCM: %s", err.message);
 		return EXIT_FAILURE;
 	}
 
-	if (argc == 2) {
-		printf("DelayAdjustment: %#.1f ms\n", (double)pcm.delay_adjustment/ 10);
+	if (argc - optind == 1) {
+		bactl_print_pcm_mute(&pcm);
 		return EXIT_SUCCESS;
 	}
 
-	const char *value = argv[optind + 1];
-	errno = 0;
-	char *endptr = NULL;
-	double dbl = strtod(value, &endptr);
-	if (endptr == value || errno == ERANGE) {
+	const char *value;
+	bool state;
+
+	value = argv[optind + 1];
+	if (!bactl_parse_value_on_off(value, &state)) {
 		cmd_print_error("Invalid argument: %s", value);
 		return EXIT_FAILURE;
 	}
 
-	int adjustment = lround(dbl * 10.0);
-	if (adjustment < INT16_MIN || adjustment > INT16_MAX) {
-		cmd_print_error("Invalid argument: %s", value);
-		return EXIT_FAILURE;
+	pcm.volume.ch1_muted = state;
+	pcm.volume.ch2_muted = state;
+
+	if (pcm.channels == 2 && argc - optind == 3) {
+
+		value = argv[optind + 2];
+		if (!bactl_parse_value_on_off(value, &state)) {
+			cmd_print_error("Invalid argument: %s", value);
+			return EXIT_FAILURE;
+		}
+
+		pcm.volume.ch2_muted = state;
+
 	}
 
-	if (!ba_dbus_pcm_set_delay_adjustment(&config.dbus, pcm.pcm_path,
-				pcm.codec.name, adjustment, &err)) {
-		cmd_print_error("DelayAdjustment update failed: %s", err.message);
+	if (!ba_dbus_pcm_update(&config.dbus, &pcm, BLUEALSA_PCM_VOLUME, &err)) {
+		cmd_print_error("Volume mute update failed: %s", err.message);
 		return EXIT_FAILURE;
 	}
 
 	return EXIT_SUCCESS;
 }
 
-const struct cli_command cmd_delay_adjustment = {
-	"delay-adjustment",
-	"Get or set PCM delay adjustment",
-	cmd_delay_adjustment_func,
+const struct bactl_command cmd_mute = {
+	"mute",
+	"Get or set PCM mute switch",
+	cmd_mute_func,
 };
