+++ src/bluealsactl/cmd-softvol.c
@@ -1,5 +1,5 @@
 /*
- * BlueALSA - cmd-info.c
+ * BlueALSA - bluealsactl/cmd-softvol.c
  * Copyright (c) 2016-2024 Arkadiusz Bokowy
  *
  * This file is a part of bluez-alsa.
@@ -9,26 +9,27 @@
  */
 
 #include <getopt.h>
+#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 
 #include <dbus/dbus.h>
 
-#include "cli.h"
+#include "bluealsactl.h"
 #include "shared/dbus-client-pcm.h"
-#include "shared/log.h"
 
 static void usage(const char *command) {
-	printf("Show PCM properties.\n\n");
-	cli_print_usage("%s [OPTION]... PCM-PATH", command);
+	printf("Get or set the SoftVolume property of the given PCM.\n\n");
+	bactl_print_usage("%s [OPTION]... PCM-PATH [STATE]", command);
 	printf("\nOptions:\n"
 			"  -h, --help\t\tShow this message and exit\n"
 			"\nPositional arguments:\n"
 			"  PCM-PATH\tBlueALSA PCM D-Bus object path\n"
+			"  STATE\t\tEnable or disable SoftVolume property\n"
 	);
 }
 
-static int cmd_info_func(int argc, char *argv[]) {
+static int cmd_softvol_func(int argc, char *argv[]) {
 
 	int opt;
 	const char *opts = "hqv";
@@ -41,7 +42,7 @@ static int cmd_info_func(int argc, char *argv[]) {
 
 	opterr = 0;
 	while ((opt = getopt_long(argc, argv, opts, longopts, NULL)) != -1) {
-		if (cli_parse_common_options(opt))
+		if (bactl_parse_common_options(opt))
 			continue;
 		switch (opt) {
 		case 'h' /* --help */ :
@@ -57,7 +58,7 @@ static int cmd_info_func(int argc, char *argv[]) {
 		cmd_print_error("Missing BlueALSA PCM path argument");
 		return EXIT_FAILURE;
 	}
-	if (argc - optind > 1) {
+	if (argc - optind > 2) {
 		cmd_print_error("Invalid number of arguments");
 		return EXIT_FAILURE;
 	}
@@ -66,20 +67,35 @@ static int cmd_info_func(int argc, char *argv[]) {
 	const char *path = argv[optind];
 
 	struct ba_pcm pcm;
-	if (!cli_get_ba_pcm(path, &pcm, &err)) {
+	if (!bactl_get_ba_pcm(path, &pcm, &err)) {
 		cmd_print_error("Couldn't get BlueALSA PCM: %s", err.message);
 		return EXIT_FAILURE;
 	}
 
-	cli_print_pcm_properties(&pcm, &err);
-	if (dbus_error_is_set(&err))
-		warn("Unable to read available codecs: %s", err.message);
+	if (argc - optind == 1) {
+		bactl_print_pcm_soft_volume(&pcm);
+		return EXIT_SUCCESS;
+	}
+
+	bool state;
+	const char *value = argv[optind + 1];
+	if (!bactl_parse_value_on_off(value, &state)) {
+		cmd_print_error("Invalid argument: %s", value);
+		return EXIT_FAILURE;
+	}
+
+	pcm.soft_volume = state;
+
+	if (!ba_dbus_pcm_update(&config.dbus, &pcm, BLUEALSA_PCM_SOFT_VOLUME, &err)) {
+		cmd_print_error("SoftVolume update failed: %s", err.message);
+		return EXIT_FAILURE;
+	}
 
 	return EXIT_SUCCESS;
 }
 
-const struct cli_command cmd_info = {
-	"info",
-	"Show PCM properties",
-	cmd_info_func,
+const struct bactl_command cmd_softvol = {
+	"soft-volume",
+	"Get or set PCM SoftVolume property",
+	cmd_softvol_func,
 };
