+++ test/test-utils-ctl.c
@@ -1,5 +1,5 @@
 /*
- * test-utils-cli.c
+ * test-utils-ctl.c
  * Copyright (c) 2016-2024 Arkadiusz Bokowy
  *
  * This file is a part of bluez-alsa.
@@ -28,10 +28,10 @@
 #include "inc/preload.inc"
 #include "inc/spawn.inc"
 
-static char bluealsa_cli_path[256];
-static int run_bluealsa_cli(char *output, size_t size, ...) {
+static char bluealsactl_path[256];
+static int run_bluealsactl(char *output, size_t size, ...) {
 
-	char * argv[32] = { bluealsa_cli_path };
+	char * argv[32] = { bluealsactl_path };
 	size_t n = 1;
 
 	va_list ap;
@@ -60,7 +60,7 @@ CK_START_TEST(test_help) {
 
 	char output[4096];
 
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"-q", "-v", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
@@ -70,7 +70,7 @@ CK_START_TEST(test_ba_service_not_running) {
 
 	char output[4096];
 
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"--dbus=test", "status", NULL), EXIT_FAILURE);
 
 } CK_END_TEST
@@ -86,12 +86,12 @@ CK_START_TEST(test_status) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"status", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check default command */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Service: org.bluealsa"), NULL);
 	ck_assert_ptr_ne(strstr(output, "A2DP-source"), NULL);
@@ -110,12 +110,12 @@ CK_START_TEST(test_list_services) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"list-services", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check service listing */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"list-services",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "org.bluealsa.test"), NULL);
@@ -136,12 +136,12 @@ CK_START_TEST(test_list_pcms) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"list-pcms", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check BlueALSA PCM listing */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"--dbus=test", "--verbose", "list-pcms",
 				NULL), 0);
 
@@ -178,17 +178,17 @@ CK_START_TEST(test_info) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"info", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check not existing BlueALSA PCM path */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"info", "/org/bluealsa/hci11/dev_FF_FF_FF_FF_FF_FF/a2dpsrc/sink",
 				NULL), EXIT_FAILURE);
 
 	/* check BlueALSA PCM info */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"info", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				"-v", "-v",
 				NULL), 0);
@@ -216,12 +216,12 @@ CK_START_TEST(test_codec) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"codec", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check BlueALSA PCM codec get/set */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"-v", "codec", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/hfpag/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Available codecs: CVSD"), NULL);
@@ -232,23 +232,23 @@ CK_START_TEST(test_codec) {
 #endif
 
 #if ENABLE_MSBC
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"codec", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/hfpag/sink", "mSBC",
 				NULL), 0);
 
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"codec", "-vf", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/hfpag/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Selected codec: mSBC"), NULL);
 #endif
 
 	/* check selecting not available codec */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"codec", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/hfpag/sink", "SBC",
 				NULL), EXIT_FAILURE);
 
 	/* check selecting A2DP codec (with our mock BlueZ) */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"codec", "-vf", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				"SBC:FF150255", "--channels=1", "--sampling=44100",
 				NULL), EXIT_SUCCESS);
@@ -268,21 +268,21 @@ CK_START_TEST(test_delay_adjustment) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"delay-adjustment", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check default delay adjustment */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"delay-adjustment", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "DelayAdjustment: 0.0 ms"), NULL);
 
 	/* check setting delay adjustment */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"delay-adjustment", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink", "-7.5",
 				NULL), 0);
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"delay-adjustment", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "DelayAdjustment: -7.5 ms"), NULL);
@@ -302,57 +302,57 @@ CK_START_TEST(test_volume) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"mute", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"soft-volume", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"volume", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check default volume */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"volume", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Volume: L: 127 R: 127"), NULL);
 
 	/* check default mute */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"mute", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Muted: L: false R: false"), NULL);
 
 	/* check default soft-volume */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"soft-volume", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "SoftVolume: true"), NULL);
 
 	/* check setting volume */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"volume", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink", "10", "50",
 				NULL), 0);
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"volume", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Volume: L: 10 R: 50"), NULL);
 
 	/* check setting mute */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"mute", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink", "off", "on",
 				NULL), 0);
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"mute", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "Muted: L: false R: true"), NULL);
 
 	/* check setting soft-volume */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"soft-volume", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink", "off",
 				NULL), 0);
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"soft-volume", "/org/bluealsa/hci11/dev_12_34_56_78_9A_BC/a2dpsrc/sink",
 				NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "SoftVolume: false"), NULL);
@@ -375,12 +375,12 @@ CK_START_TEST(test_monitor) {
 	char output[4096];
 
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"monitor", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
 	/* check monitor command */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"monitor", "-v", "--properties=codec,volume",
 				NULL), 0);
 
@@ -428,41 +428,41 @@ CK_START_TEST(test_open) {
 
 	char output[4096];
 	/* check printing help text */
-	ck_assert_int_eq(run_bluealsa_cli(output, sizeof(output),
+	ck_assert_int_eq(run_bluealsactl(output, sizeof(output),
 				"open", "--help", NULL), 0);
 	ck_assert_ptr_ne(strstr(output, "-h, --help"), NULL);
 
-	char * ba_cli_in_argv[32] = {
-		bluealsa_cli_path, "open", "--hex",
+	char * bactl_in_argv[32] = {
+		bluealsactl_path, "open", "--hex",
 		"/org/bluealsa/hci11/dev_23_45_67_89_AB_CD/hspag/source",
 		NULL };
-	char * ba_cli_out_argv[32] = {
-		bluealsa_cli_path, "open", "--hex",
+	char * bactl_out_argv[32] = {
+		bluealsactl_path, "open", "--hex",
 		"/org/bluealsa/hci11/dev_23_45_67_89_AB_CD/hspag/sink",
 		NULL };
 
-	struct spawn_process sp_ba_cli_in;
-	ck_assert_int_ne(spawn(&sp_ba_cli_in, ba_cli_in_argv,
+	struct spawn_process sp_bactl_in;
+	ck_assert_int_ne(spawn(&sp_bactl_in, bactl_in_argv,
 				NULL, SPAWN_FLAG_REDIRECT_STDOUT), -1);
 
-	struct spawn_process sp_ba_cli_out;
-	ck_assert_int_ne(spawn(&sp_ba_cli_out, ba_cli_out_argv,
-				sp_ba_cli_in.f_stdout, SPAWN_FLAG_NONE), -1);
+	struct spawn_process sp_bactl_out;
+	ck_assert_int_ne(spawn(&sp_bactl_out, bactl_out_argv,
+				sp_bactl_in.f_stdout, SPAWN_FLAG_NONE), -1);
 
 	/* let it run for a while */
 	usleep(250000);
 
-	spawn_terminate(&sp_ba_cli_in, 0);
-	spawn_terminate(&sp_ba_cli_out, 500);
+	spawn_terminate(&sp_bactl_in, 0);
+	spawn_terminate(&sp_bactl_out, 500);
 
 	int wstatus = 0;
-	/* Make sure that input bluealsa-cli instances have been terminated by
+	/* Make sure that input bluealsactl instances have been terminated by
 	 * us (SIGTERM) and not by premature exit or any other reason. On the other
-	 * hand, the output bluealsa-cli instance should exit gracefully because
+	 * hand, the output bluealsactl instance should exit gracefully because
 	 * of the end of input stream. */
-	spawn_close(&sp_ba_cli_in, &wstatus);
+	spawn_close(&sp_bactl_in, &wstatus);
 	ck_assert_int_eq(WTERMSIG(wstatus), SIGTERM);
-	spawn_close(&sp_ba_cli_out, &wstatus);
+	spawn_close(&sp_bactl_out, &wstatus);
 	ck_assert_int_eq(WIFEXITED(wstatus), 1);
 	ck_assert_int_eq(WEXITSTATUS(wstatus), 0);
 
@@ -479,8 +479,8 @@ int main(int argc, char *argv[]) {
 
 	snprintf(bluealsa_mock_path, sizeof(bluealsa_mock_path),
 			"%s/mock/bluealsa-mock", argv_0_dir);
-	snprintf(bluealsa_cli_path, sizeof(bluealsa_cli_path),
-			"%s/../utils/cli/bluealsa-cli", argv_0_dir);
+	snprintf(bluealsactl_path, sizeof(bluealsactl_path),
+			"%s/../src/bluealsactl/bluealsactl", argv_0_dir);
 
 	Suite *s = suite_create(__FILE__);
 	TCase *tc = tcase_create(__FILE__);
